+++ test/completion/import_tree/recurse_class1.py
@@ -1,4 +1,4 @@
-import recurse_class2
+from . import recurse_class2
 
 class C(recurse_class2.C):
     def a(self):