+++ CHANGELOG.rst
@@ -3,6 +3,12 @@
 Changelog
 ---------
 
+0.14.2 (XXXX-XX-XX)
++++++++++++++++++++
+
+- Better support for enums/dataclasses
+
+
 0.14.1 (2019-07-13)
 +++++++++++++++++++